make dd comon slimmer via features
