+++ ddcommon/Cargo.toml
@@ -11,23 +11,29 @@ version = "2.0.0"
 crate-type = ["lib"]
 
 [dependencies]
-anyhow = "1.0"
-futures = "0.3"
+anyhow = {version = "1.0"}
+futures = {version = "0.3"}
 futures-core = {version = "0.3.0", default-features = false}
 futures-util = {version = "0.3.0", default-features = false}
-hex = "0.4"
-hyper = {version = "0.14", features = ["http1", "client", "tcp", "stream"], default-features = false}
-hyper-rustls = {version = "0.23", default-features = false, features = ["native-tokio", "http1", "tls12"]}
-lazy_static = "1.4"
+hex = {version = "0.4"}
+hyper = {version = "0.14", features = ["http1", "client", "tcp", "stream"], default-features = false, optional = true}
+hyper-rustls = {version = "0.23", default-features = false, features = ["native-tokio", "http1", "tls12"], optional = true}
+lazy_static = {version = "1.4"}
 log = {version = "0.4"}
-pin-project = "1"
-regex = "1.5"
-rustls = {version = "0.20.4", default-features = false}
-rustls-native-certs = {version = "0.6"}
+pin-project = {version = "1", optional = true}
+regex = {version = "1.5"}
+rustls = {version = "0.20.4", default-features = false, optional = true}
+rustls-native-certs = {version = "0.6", optional = true}
 tokio = {version = "1.23", features = ["rt", "macros"]}
 tokio-rustls = {version = "0.23"}
 serde = {version = "1.0", features = ["derive"]}
 
 [dev-dependencies]
-indexmap = "1.8"
-maplit = "1.0"
+indexmap = {version="1.8"}
+maplit = {version="1.0"}
+
+[features]
+default = ["agent-connector", "agent-http"]
+
+agent-connector = ["hyper", "hyper-rustls", "rustls", "rustls-native-certs", "pin-project"]
+agent-http = ["hyper"]
